reinstalling backend
